Upload existing public key
We need to do this rather than generate new keys each time since we're running simply as a proxy command and can't inject additional command line parameters like `-i` to use the appropriate key.

Also just truncate the entire `authorized_keys` file on the server since backslash escaping hell to remove the key is inevitably going to break one day and cause the keys to be left around. If people ssh into the same machine in parallel it will potentially break it for the other person but so be it.